+++ js/util.js
@@ -1,0 +1,20 @@
+function getRandomInteger(a, b) {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+const getRandomArrayElement = (array) =>
+  array[getRandomInteger(0, array.length - 1)];
+
+const createIdgenerator = () => {
+  let lastgenerateId = 0;
+
+  return () => {
+    lastgenerateId += 1;
+    return lastgenerateId;
+  };
+};
+
+export {getRandomInteger, getRandomArrayElement, createIdgenerator};
